Updated unit tests
-Added tests for functions
-Added error class and return from package manager functions
-Updated `compile.sh` script

+++ bin/compile.sh
@@ -9,9 +9,15 @@
 # CMake/ninja build system
 mkdir -p build
 cmake -B build -S . -DCMAKE_EXPORT_COMPILE_COMMANDS=1 -G Ninja
-ninja -C build
+ninja -C build -j $(nproc)
 
 # Create symlinks to executables in build folder if necessary
-rm bin/gdpm bin/gdpm-tests
-ln -s ../build/gdpm bin/gdpm
-ln -s ../build/gdpm-tests bin/gdpm-tests
+if test -f "../build/gdpm"; then
+	rm bin/gdpm
+	ln -s ../build/gdpm bin/gdpm
+fi
+
+if test -f "../build/gdpm-tests"; then
+	rm bin/gdpm-tests
+	ln -s ../build/gdpm-tests bin/gdpm-tests
+fi
